refactor: RegameChoice enum을 Controller와 Model로 분리
이전의 RegameChoice enum이 Model과 Controller 2가지 역할을 하고 있다고 판단.
사용자가 입력한 값을 검증하는 기능인 validateValue method는 Controller에 담음.
사용자가 입력할 수 있는 값들은 enum에 담음.
이름을 RegameChoice에서 Regame으로 단순화하여 가독성 향상, 통일성 향상.
