--- src/main/java/baseball/controller/RegameChoice.java
@@ -1,27 +1,0 @@
-package baseball.controller;
-
-public enum RegameChoice {
-    RESTART_GAME(1),
-    EXIT_GAME(2),
-    OTHER_CHOICE(3);
-
-    private final int value;
-
-    RegameChoice(int value) {
-        this.value = value;
-    }
-
-    public int getValue() {
-        return value;
-    }
-
-    public static RegameChoice fromValue(int value) {
-        for (RegameChoice choice : RegameChoice.values()) {
-            if (choice.value == value) {
-                return choice;
-            }
-        }
-        return OTHER_CHOICE;
-    }
-}
-
